Add xk6-redis as experimental redis js module
This commit imports the latest state of xk6-redis into k6's core as
an experimental module importable as "k6/experimental/redis".

With it comes a single direct dependency, the `redis/v8` library, and
two indirect dependencies implied by the former `xxhash`
and `go-rendezvous`.

This commit's essential files are:
* `js/modules/k6/experimental/redis`: the files composing the module
itself
* `js/initcontext.go`: exposes the module as `k6/experimental/redis`
* `samples/redis.go`: presents an example of the redis module in
action

The rest of the files are essentially go module and vendoring plumbing.